Refactor snake movement and boundary calculations
Refactored the movement and boundary calculations for the snake object. Introduced a new constant, EDGE_POSITION, to account for grid size when defining boundaries. This prevents the snake from disappearing at the left and top window edges. The code was further modified to provide a Position struct which hones in on the concept of positions rather than using separate x and y properties. Simplified the syntax of direction op-positiveness check to make it more readable. Adjusted the calculation of new piece's position to use a grid-aligned position within the specified window size. This modification encapsulates grid calculations and window size verification. Improved RNG in `get_random_coordinate` by passing it as an argument to make it more efficient. The graphic transformation code was also compressed to streamline appearance rendering. These changes increase overall code readability and efficiency.